requirements.txt
flask==1.1.2
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
pytest==6.2.3
gunicorn==20.0.4
python-dotenv==0.14.0
<<<<<<< Updated upstream
httpie==2.4.0
=======
httpie==2.4.0
pytest-cov==2.10.1
flake8==3.8.4
black==20.8b1
isort==5.6.4
>>>>>>> Stashed changes
